+++ vscode/src/local-context/download-symf.ts
@@ -10,7 +10,7 @@ import { captureException } from '../services/sentry/sentry'
 import { downloadFile, fileExists, unzip } from './utils'
 
 export type SymfVersionString = SemverString<'v'>
-const symfVersion: SymfVersionString = 'v0.0.13'
+const symfVersion: SymfVersionString = 'v0.0.16'
 
 export const _config = {
     //delay before trying to re-lock a active file